# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ecommerce-BackEnd-Documentation",
        "description": "Docuementation for Ecommerce-BackEnd",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8017",
            "description": "Develop Server"
        }
    ],
    "paths": {
...
        "/api/media/getmedia": {
            "get": {
                "tags": [
                    "Media"
                ],
                "summary": "Obtener todos los Medios",
                "description": "Este EndPoint regresa todos los Medios.",
                "operationId": "allMedia",
                "responses": {
                    "200": {
                        "description": "Todos los Medios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "phone_number": {
                                                        "type": "string",
                                                        "example": "1234567899"
                                                    },
                                                    "skype": {
                                                        "type": "string",
                                                        "example": "example.skype"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2024-12-12"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2024-12-12"
                                                    },
                                                    "deleted_at": {
                                                        "type": "string",
                                                        "example": "null"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Medios obtenidos correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No autorizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error del servidor"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/media/getMedia/{id}": {
            "get": {
                "tags": [
                    "Media"
                ],
                "summary": "Obtener un Medio",
                "description": "Este EndPoint regresa un Medio por su ID.",
                "operationId": "getmedia",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del Medio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Medio obtenido correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "phone_number": {
                                                    "type": "string",
                                                    "example": "1234567899"
                                                },
                                                "skype": {
                                                    "type": "string",
                                                    "example": "example.skype"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-12-12"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-12-12"
                                                },
                                                "deleted_at": {
                                                    "type": "string",
                                                    "example": "null"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Medio obtenido correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No autorizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Medio no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Medio no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error del servidor"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/media/deleteMedia/{id}": {
            "delete": {
                "tags": [
                    "Media"
                ],
                "summary": "Eliminar medio.",
                "description": "Este endpoint permite eliminar al medio.",
                "operationId": "deleteMedia",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del Medio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Medio eliminado correctamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Medio eliminado correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No autorizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lo sentimos, ha ocurrido un error interno en el servidor."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}